+++ client/src/views/NeedFormVolunteers.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button, Grid } from '@mui/material';
 
 import { RadioGroup, Select, TextField } from '../components/Forms';
-import NeedOfferForm from './NeedOfferForm';
+import NeedOfferForm from '../components/Forms/NeedOfferForm';
 
 const skillCategories = [
   { value: 'purple', text: 'Purple' },
